refactor things
