+++ packages/app/app/store/message/room/call.ts
@@ -223,8 +223,8 @@ export const useCallStore = defineStore("message/room/call", () => {
     currentRoomCallSessionId,
     deleteCallParticipant,
     deleteSpeaker,
-    isDeafened,
     isCameraEnabled,
+    isDeafened,
     isForceMuted,
     isInCall,
     isMuted,
