+++ packages/app/server/models/room/call/JoinCallOutput.ts
@@ -1,0 +1,8 @@
+import type { CallParticipant } from "#shared/models/room/call/CallParticipant";
+
+export interface JoinCallOutput {
+  callSessionId: string;
+  livekitToken: string;
+  livekitUrl: string;
+  participants: CallParticipant[];
+}
