+++ packages/app/server/services/message/call/createParticipant.ts
@@ -1,0 +1,13 @@
+import type { CallParticipant } from "#shared/models/room/call/CallParticipant";
+
+export const createParticipant = (
+  session: { id: string },
+  user: { id: string; image?: null | string; name: string },
+): CallParticipant => ({
+  id: session.id,
+  image: user.image ?? null,
+  isCameraEnabled: false,
+  isMuted: false,
+  name: user.name,
+  userId: user.id,
+});
