+++ packages/app/server/services/message/call/deleteCallParticipant.ts
@@ -4,7 +4,7 @@ export const deleteCallParticipant = (callSessionId: string, id: string): boolea
   const participantMap = callSessionParticipantMap.get(callSessionId);
   if (!participantMap) return false;
 
-  const deletedParticipant = participantMap.delete(id);
+  const isDeleted = participantMap.delete(id);
   if (participantMap.size === 0) callSessionParticipantMap.delete(callSessionId);
-  return deletedParticipant;
+  return isDeleted;
 };
